added a jotai globla state component to allow the rendering of the patient creation modal from anywhere
