+++ hooks/patients.ts
@@ -43,6 +43,10 @@ export default function usePatients(pagination?: UserPagination): UsePatients {
     data?.patients.filter((patient) => patient.pickupEnabled && patient.dob) ??
     null
 
+  async function addPatient(){
+    
+  }
+
   return {
     patients: data?.patients ?? null,
     numPatients: data?.numPatients ?? null,
